[libromdata] DreamcastSave, WiiBNR: Reduced the type sizes in some constexpr arrays.
Code size difference: (64-bit Gentoo Linux, gcc-13.2.1_p20240210, release build, no LTO)

   text    data     bss     dec     hex filename
   3551     344       0    3895     f37 WiiBNR.cpp.o [before]
   3545     344       0    3889     f31 WiiBNR.cpp.o [after]
     -6       0       0      -6      -6 Difference

   text    data     bss     dec     hex filename
  19869     496       0   20365    4f8d DreamcastSave.cpp.o
  19861     496       0   20357    4f85 DreamcastSave.cpp.o
     -8       0       0      -8      -8 Difference

+++ src/libromdata/Console/DreamcastSave.cpp
@@ -130,7 +130,7 @@ public:
 	int readVmiHeader(const IRpFilePtr &vmi_file);
 
 	// Graphic eyecatch sizes (in bytes)
-	static const array<uint32_t, 4> eyecatch_sizes;
+	static const array<uint16_t, 4> eyecatch_sizes;
 
 	// VMS icon struct
 	// For processing VMS icons only;
@@ -209,7 +209,7 @@ const RomDataInfo DreamcastSavePrivate::romDataInfo = {
 };
 
 // Graphic eyecatch sizes (in bytes)
-const array<uint32_t, 4> DreamcastSavePrivate::eyecatch_sizes = {{
+const array<uint16_t, 4> DreamcastSavePrivate::eyecatch_sizes = {{
 	0,	// DC_VMS_EYECATCH_NONE
 	DC_VMS_EYECATCH_ARGB4444_DATA_SIZE,
 	DC_VMS_EYECATCH_CI8_PALETTE_SIZE + DC_VMS_EYECATCH_CI8_DATA_SIZE,
